% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipShape.R
\name{clipShape}
\alias{clipShape}
\title{Clip supplied shapefiles}
\usage{
clipShape(xmin, xmax, ymin, ymax)
}
\arguments{
\item{xmin}{numeric minimum latitude}

\item{xmax}{numeric maximum latitude}

\item{ymin}{numeric minimum longitude}

\item{ymax}{numeric maximum longitude}
}
\value{
list of politicalBoundary, rivers, and lakes shapefiles, and a vector of lat/lon limits
}
\description{
Clips the political boundary, hydrolines, and hydroshape shapefiles to user-defined lat/lon limits.
}
\examples{
xmin <- -88-11/60 #MMSD
xmax <- -87-48/60 #MMSD
ymin <- 42+51/60 #MMSD
ymax <- 43+20/60 #MMSD
retList <- clipShape(xmin,xmax,ymin,ymax)
polShape <- retList$politicalBoundary
riverShape <- retList$rivers
lakeShape <- retList$lakes
}

