% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/packageInfo.R
\docType{data}
\name{dfOptical}
\alias{dfOptical}
\title{Example data for correlations}
\description{
Example data with response and predictor variables
}
\author{
Steve Corsi \email{srcorsi@usgs.gov}
}
\keyword{correlation}

