% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/packageInfo.R
\docType{data}
\name{events}
\alias{events}
\title{Example event dates and times.}
\description{
Example event begin and end dates and times to define a sampled hydrograph
}
\author{
Steve Corsi \email{srcorsi@usgs.gov}
}
\keyword{dates}
\keyword{event}
\keyword{hydrograph}

