% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageInfo.R
\docType{data}
\name{exampleHYSEP}
\alias{exampleHYSEP}
\title{Example HYSEP output data.}
\description{
Example HYSEP output data. Needs more info.
}
\author{
Steve Corsi \email{srcorsi@usgs.gov}
}
\keyword{instantaneous}

