% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageInfo.R
\docType{data}
\name{FIBdata}
\alias{FIBdata}
\title{Example water quality data included in GSHydroTools}
\description{
Example data representing composite fecal indicator bacteria from the Menomonee River at Wauwatosa, Wisconsin
}
\author{
Steve Corsi \email{srcorsi@usgs.gov}
}
\keyword{bacteria}
\keyword{data}

