% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/packageInfo.R
\docType{data}
\name{flowData}
\alias{flowData}
\title{Example flow data.}
\description{
Example instantaneous (unit value) flow data from Menomonee River at Wauwatosa, WI.
}
\author{
Steve Corsi \email{srcorsi@usgs.gov}
}
\keyword{instantaneous}

