% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/packageInfo.R
\docType{package}
\name{GSHydroTools-package}
\alias{GSHydroTools-package}
\title{Collection of functions for hydrological analysis.}
\description{
\tabular{ll}{
Package: \tab GSHydroTools\cr
Type: \tab Package\cr
Version: \tab 1.0.0\cr
Date: \tab 2014-01-10\cr
License: \tab Unlimited for this package, dependencies have more restrictive licensing.\cr
Copyright: \tab This software is in the public domain because it contains materials
that originally came from the United States Geological Survey, an agency of
the United States Department of Interior. For more information, see the
official USGS copyright policy at
http://www.usgs.gov/visual-id/credit_usgs.html#copyright\cr
LazyLoad: \tab yes\cr
}
}
\details{
Collection of functions for hydrological analysis.
}
\author{
Steve Corsi \email{srcorsi@usgs.gov}
}

