% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiCor.R
\name{multiCor}
\alias{multiCor}
\title{multiCor 
compute correlation coefficients for one response variable vs multiple predictor (independent) variables. 
The output is a dataframe ordered by highest to lowest correlation}
\usage{
multiCor(df, response, IVs, method = "spearman")
}
\arguments{
\item{df}{is dataframe with response variable and predictor variables}

\item{response}{is a character string that is the name of the response variable in df}

\item{IVs}{is a vector of character strings that are the independent variables in df}

\item{method}{is either "spearman" (nonparametric) or "pearson" (parametric)}
}
\value{
z dataframe with the variable name in column 1 and correlation 
coefficient in column 2. The dataframe is ordered from greatest correlation
to least correlation.
}
\description{
multiCor 
compute correlation coefficients for one response variable vs multiple predictor (independent) variables. 
The output is a dataframe ordered by highest to lowest correlation
}
\examples{
data <- dfOptical
multiCor(data,"logEColi",names(data)[-1],"spearman")
}

