% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotBackgroundMap.R
\name{plotBackgroundMap}
\alias{plotBackgroundMap}
\title{Set up map with political, lake, and river boundaries.}
\usage{
plotBackgroundMap(shapeList)
}
\arguments{
\item{shapeList}{list returned from clipShape function (includes 3 shapefiles and a vector of lat/lon limits)}
}
\value{
list of politicalBoundary, rivers, and lakes shapefiles
}
\description{
Set up map with political, lake, and river boundaries.
}
\examples{
xmin <- -88-11/60 #MMSD
xmax <- -87-48/60 #MMSD
ymin <- 42+51/60 #MMSD
ymax <- 43+20/60 #MMSD
retList <- clipShape(xmin,xmax,ymin,ymax)
plotBackgroundMap(retList)
}

