% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHydroConc.R
\name{plotHydroConc}
\alias{plotHydroConc}
\title{plotHydroConc}
\usage{
plotHydroConc(Q, QVars, QDateVars, smooth, sites, Conc, CVars, CDateVar,
  CVarsDisplay, dates, sampDates, eventDates, Qcols, Ccols, leftBuffer,
  rightBuffer, concLines = TRUE, Qylab = "Discharge (cfs)",
  Cylab = "Concentration", title1 = "Flow", title2 = "and concentration")
}
\arguments{
\item{Q}{dataframe list with flow variables and POSIXct date variables. The 
list contains one file per Q record (usually one dataframe per site).}

\item{QVars}{vector of strings that signify the column names that represent 
the flow variables in the dataframes defined in Q}

\item{QDateVars}{vector of strings that signify the column names that represent 
the POSIXct date variables in the dataframes defined in Q}

\item{smooth}{Boolean vector to trigger lowess smooth in graphing rather than
direct flow graphing (useful for sites impacted by seiche)}

\item{sites}{Sites for Q data. This will be used in the legend of the Q panel}

\item{Conc}{dataframe with variables to be plotted in the second panel}

\item{CVars}{vector of strings that represent the variables in Conc to include
in the graphs in the second panel}

\item{CDateVar}{column name in Conc for sample dates and times in POSIXct
#'@param CVarsDisplay variable names from CVars to be displayed on the legend}

\item{dates}{dataframe with beginning and ending sample dates and times 
and beginning and ending hydrograph dates and times}

\item{sampDates}{vector of length two that contains strings representing
the column names for beginning and ending dates and times for samples}

\item{eventDates}{vector of length two that contains strings representing
the column names for beginning and ending dates and times for the sampling 
event. This is the hydrograph segment that the samples are intended to represent
(often a runoff hydrograph or a baseflow period)}

\item{Qcols}{vector of colors for ploting hydrographs from the Q dataframe}

\item{Ccols}{vector of colors for plotting from the Conc dataframe}

\item{leftBuffer}{time in days to plot before the beginning of the event period}

\item{rightBuffer}{time in days to plot after the end of the event period}

\item{concLines}{Boolean variable to signify whether a line should be drawn 
between consecutive data points from the Conc dataframe in the second graph 
panel}

\item{Qylab}{y-axis label for the first graph panel}

\item{Cylab}{y-axis label for the second graph panel}

\item{title1}{Line 1 of plot title}

\item{title2}{Line 2 of plot title}
}
\description{
Function to generate a two panel graph with hydrograph(s) in the 
top panel and concentration or other water quality parameter (e.g. flux) 
in the lower panel with corresponding date axes for the top and bottom panels
}
\examples{
#Add example

}

