% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageInfo.R
\docType{data}
\name{shape_hydroline}
\alias{shape_hydroline}
\title{Rivers in US.}
\description{
River shapefiles from \url{http://dds.cr.usgs.gov/pub/data/nationalatlas/hydro0m_shp_nt00300.tar.gz}
}
\keyword{data}
\keyword{station}

