% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{shape_polibounds}
\alias{shape_polibounds}
\title{Political boundaries.}
\description{
Political boundary shapefiles from \url{http://dds.cr.usgs.gov/pub/data/nationalatlas/bound0m_shp_nt00298.tar.gz}
}
\keyword{data}
\keyword{station}

