\docType{data}
\name{subShape_hydropolyClip}
\alias{subShape_hydropolyClip}
\title{Lakes near and including the Great Lakes.}
\description{
Great Lakes shapefiles from
\url{http://dds.cr.usgs.gov/pub/data/nationalatlas/bound0m_shp_nt00298.tar.gz}
cropped to the Great Lakes Region
}
\keyword{data}
\keyword{station}

