% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeData.R
\name{summarizedataDF}
\alias{summarizedataDF}
\title{Summarize dataDF by group}
\usage{
summarizedataDF(dataDF, colGroup, colValue, colDate)
}
\arguments{
\item{dataDF}{dataframe with columns defined by colGroup (grouping column, such as site ID), colValue (value column),
and optionally colDate (date columns)}

\item{colGroup}{string defines grouping column in dataDF}

\item{colValue}{string defines value column in dataDF}

\item{colDate}{string defines date column in dataDF. If colDate = NA, the calculations for start and end date are ignored.}
}
\value{
dataframe with count, mean, median, min, max, start(date/time), end(date/time), and number of non-detects (nd) defined as number of NA's grouped by colGroup
}
\description{
Mapping routine that displays spatial dataDF variability by color differences.
over layers with political boundaries, hydrologic polygons, and hydrologic lines.
}
\examples{
df <- data.frame(site=c("1","x","2","1","x","2"),
                 conc=c(2,3,4,5,NA,7),
                 dates=as.Date(c("2011-01-01","2011-01-01",
                   "2011-01-01","2011-01-02","2011-01-02","2011-01-02")))
sumDF <- summarizedataDF(df, "site", "conc", "dates")
}

