\docType{data}
\name{WQdata}
\alias{WQdata}
\title{Example water quality data included in GSHydroTools}
\description{
  Example data representing discrete total phosphorus
  concentrtions from the Menomonee River at Wauwatosa,
  Wisconsin
}
\author{
  Steve Corsi \email{srcorsi@usgs.gov}
}
\keyword{data}
\keyword{datasets}
\keyword{quality}
\keyword{water}

