% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fxn_MapColor.R
\name{MapColor}
\alias{MapColor}
\title{Mapping data with variable color to define spatial data}
\usage{
MapColor(df, colorVar, latVar, lonVar, xmin, xmax, ymin, ymax, col1 = "tan",
  col2 = "orange3", col3 = "orangered1", col4 = "orangered4", xleft,
  xright, ytop, ybottom, mainTitle = "", units = 6, includeLabels = FALSE,
  labels = "", offsetLat = "", offsetLon = "", offsetLineLat = "",
  offsetLineLon = "", DL = 0, LegCex = 0.9, titlePos = -4)
}
\arguments{
\item{df}{Dataframe with 3 required columns to define symbol color,
latitude, and longitude of points to be plotted, and optionally 5 addtional columns: label names,
offset for latitude and longitude for labels, and offset of latitude and longitude for the ending}

\item{colorVar}{Column name in df to define symbol color}

\item{latVar}{Column name in df to define latitude}

\item{lonVar}{Column name in df to define longitude}

\item{xmin}{Left longitudinal boundary for plotting}

\item{xmax}{Right longitudinal boundary for plotting}

\item{ymin}{Bottom latitudinal boundary for plotting}

\item{ymax}{Top latitudinal boundary for plotting}

\item{col1}{Symbol color for 0-0.25 quantile bin}

\item{col2}{Symbol color for 0.25-0.5 quantile bin}

\item{col3}{Symbol color for 0.5-0.75 quantile bin}

\item{col4}{Symbol color for 0.75-1.0 quantile bin}

\item{xleft}{Placement of left side of legend box (min latitude)}

\item{xright}{Placement of right side of legend box (min latitude)}

\item{ytop}{Placement of top side of legend box (min longitude)}

\item{ybottom}{Placement of bottom side of legend box (min longitude)}

\item{mainTitle}{Text to be used as the title of the plot}

\item{units}{1-4 are for water concentration: mg/L, ug/L, ng/L, pg/L respectively.
5-6 are for sediment concentration: mg/kg, ug/kg, ng/kg, pg/kg respectively.}

\item{includeLabels}{logical, if TRUE labels will be included on plot.}

\item{labels}{String variable in dataframe df with label names}

\item{offsetLat}{Variable in dataframe df for the offset from dataLat used
for label positioning}

\item{offsetLon}{Variable in dataframe df for the offset from dataLon used
for label positioning}

\item{offsetLineLat}{Variable in dataframe df for the offset from dataLat used to
position the end of the line drawn to the label. Lines are optional.}

\item{offsetLineLon}{Variable in dataframe df for the offset from dataLon used to
position the end of the line drawn to the label. Lines are optional.}

\item{DL}{numeric vector of detection limits}

\item{LegCex}{size of the text and symbols in the legend as numeric. Assigns the
pt.cex and cex arguments in legend() and text().}

\item{titlePos}{position of title as numeric. Assigns the line() argument in mtext().}
}
\description{
Mapping routine that displays spatial data variability by color differences.
over layers with political boundaries, hydrologic polygons, and hydrologic lines.
}
\examples{
lat <- SI$lat
lon <-  SI$lon
y <- runif(n=length(lat),min=0,max=50)
df <- data.frame(y=y, lat=lat,lon=lon)
colorVar <- "y"
latVar <- "lat"
lonVar <- "lon"
DL <- c(rep(0.05,times=19),rep(0.04,times=10))
LegCex <- 0.7
titlePos <- -4

xmin <- -96.5
xmax <- -72
ymin <- 40.5
ymax <- 49.5
xleft <- -95
ybottom <- 40.7
xright <- -90.8
ytop <- 43.5
mainTitle <- "OC Pesticides"

#Without labels

#Example works best in a landscape view:
pdf("GreatLakesExamplePlotNoLabels.pdf",width=11,height=8)
MapColor(df,colorVar, latVar, lonVar,
         xmin,xmax,ymin,ymax,
         xleft=xleft, xright=xright, ytop=ytop, ybottom=ybottom,
         mainTitle=mainTitle,
         includeLabels=FALSE, DL=DL, LegCex=LegCex, titlePos=titlePos)
dev.off()
#To view the produced plot, us the following command:
\dontrun{shell.exec("GreatLakesExamplePlotNoLabels.pdf")}

# With labels:
df <- merge(df, SI, by=c("lat","lon"))
labelVar <- "Site"
offsetLatVar <- "offsetLat"
offsetLonVar <- "offsetLon"
offsetLineLatVar <- "offsetLineLat"
offsetLineLonVar <- "offsetLineLon"
#Example works best in a landscape view:
pdf("GreatLakesExamplePlot.pdf",width=11,height=8)
MapColor(df,colorVar,latVar,lonVar,
         xmin,xmax,ymin,ymax,
         xleft=xleft,xright=xright,ytop=ytop,ybottom=ybottom,
         mainTitle=mainTitle, includeLabels=TRUE,
         labels=labelVar, offsetLat=offsetLatVar,
         offsetLon=offsetLonVar, offsetLineLat=offsetLineLatVar,
         offsetLineLon=offsetLineLonVar, DL=DL, LegCex=LegCex, titlePos=titlePos)
dev.off()
#To view the produced plot, us the following command:
\dontrun{shell.exec("GreatLakesExamplePlot.pdf")}
}
\keyword{color}
\keyword{map}
\keyword{spatial}

