% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{subShape_hydropolyClip}
\alias{subShape_hydropolyClip}
\title{Lakes near and including the Great Lakes.}
\description{
Great Lakes shapefiles from \url{http://dds.cr.usgs.gov/pub/data/nationalatlas/hydro0m_shp_nt00300.tar.gz} cropped to the Great Lakes Region
}
\keyword{data}
\keyword{station}

