% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WQcompos.R
\name{WQcompos}
\alias{WQcompos}
\title{WQcompos}
\usage{
WQcompos(df.samples, sampleID, parms, volume = "Evolume", bdate, edate, codes)
}
\arguments{
\item{df.samples}{dataframe with sample results and volumes}

\item{sampleID}{character variable name for the IDs for compositing samples (multiple samples will have the same ID)}

\item{parms}{vector Parameters to composite}

\item{volume}{character variable name for the volume, defaults to "Evolume"}

\item{codes}{a vector of character variable names for the values that should be pasted together into one string when combining samples (lab IDs are common here)}

\item{btime}{character variable name for the beginning of event times for each sample}

\item{etime}{character variable name for the ending of event times for each sample}
}
\value{
IDdf dataframe
}
\description{
Combine multiple samples into one composite sample (loads and concentrations)
}
\details{
function to composite samples weighted by the associated volume
the result is a volume-weighted concentration and summation of volumes
}
\examples{
flowData <- flowData
FIBdata <- FIBdata
FIBcomposData <- Hydrovol(dfQ=flowData,Q="Q",time="pdate",df.dates=FIBdata,bdate="SSdate",edate="SEdate")
WQcompos(df.samples=FIBcomposData,sampleID="SampleID",parms=c("Ecoli","Enterococci"),volume="event.vol",bdate="SSdate",edate="SEdate",codes="SampleID")
}

